% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.data.meal.R
\docType{data}
\name{fv.data.meal}
\alias{fv.data.meal}
\title{Test Data for `fvs_scores_meal`}
\format{
`fv.data.meal`
\describe{
A data frame with 45 rows and 30 columns:
  \item{Q1}{Fruit juice frequency of consumption}
  \item{Q1A}{Fruit juice portion size}
  \item{Q2}{Salad or lettuce frequency of consumption}
  \item{Q2A}{Salad or lettuce portion size}
  \item{Q3}{French fry frequency of consumption}
  \item{Q3A}{French fry portion size}
  \item{Q4}{Other potatoes frequency of consumption}
  \item{Q4A}{Other potatoes portion size}
  \item{Q5}{Beans frequency of consumption}
  \item{Q5A}{Beans portion size}
  \item{Q6}{Fruit AM frequency of consumption}
  \item{Q6A}{Fruit AM portion size}
  \item{Q7}{Other vegetables AM frequency of consumption}
  \item{Q7A}{Other vegetables AM portion size}
  \item{Q8}{Fruit NN frequency of consumption}
  \item{Q8A}{Fruit NN portion size}
  \item{Q8}{Tomato sauce frequency of consumption}
  \item{Q8A}{Tomato sauce portion size}
  \item{Q9}{Other vegetables NN frequency of consumption}
  \item{Q9A}{Other vegetables NN portion size}
  \item{Q10}{Fruit PM frequency of consumption}
  \item{Q10A}{Fruit PM portion size}
  \item{Q11}{Other vegetables PM frequency of consumption}
  \item{Q11A}{Other vegetables PM portion size}
  \item{Q12}{Tomato sauce frequency of consumption}
  \item{Q12A}{Tomato sauce portion size}
  \item{Q13}{Vegetable souop frequency of consumption}
  \item{Q13A}{Vegetable soup portion size}
  \item{Q14}{Mixtures with vegetables frequency of consumption}
}
}
\source{
<https://epi.grants.cancer.gov/diet/screeners/fruitveg/scoring/bymeal.html>
}
\usage{
fv.data.meal
}
\description{
Test data for examples in the `fvs_scores_meal` function.
}
\keyword{datasets}
