% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.data.R
\docType{data}
\name{fv.data}
\alias{fv.data}
\title{Test Data for fvs_scores_day}
\format{
`fv.data`
\describe{
A data frame with 45 rows and 20 columns:
  \item{Q1}{Fruit juice frequency of consumption}
  \item{Q1A}{Fruit juice portion size}
  \item{Q2}{Fruit}
  \item{Q2A1}{Fruit portion size option A}
  \item{Q2A2}{Fruit portion size option B}
  \item{Q3}{Salad or lettuce frequency of consumption}
  \item{Q3A}{Salad or lettuce portion size}
  \item{Q4}{French fry frequency of consumption}
  \item{Q4A}{French fry portion size}
  \item{Q5}{Other potatoes frequency of consumption}
  \item{Q5A}{Other potatoes portion size}
  \item{Q6}{Beansfrequency of consumption}
  \item{Q6A}{Beans portion size}
  \item{Q7}{Other vegetanles frequency of consumption}
  \item{Q7A}{Other vegetables portion size}
  \item{Q8}{Tomato sauce frequency of consumption}
  \item{Q8A}{Tomato sauce portion size}
  \item{Q9}{Vegetable souop frequency of consumption}
  \item{Q9A}{Vegetable soup portion size}
  \item{Q10}{Mixtures with vegetables frequency of consumption}
}
}
\source{
<https://epi.grants.cancer.gov/diet/screeners/fruitveg/scoring/allday.html>
}
\usage{
fv.data
}
\description{
Test data for examples in the fvs_scores_day function.
}
\keyword{datasets}
