% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvs_scores_day.R
\name{fvs_scores_day}
\alias{fvs_scores_day}
\title{Scores for the Fruit & Vegetable Intake Day Screener in the Eating at America's Table Study (EATS)}
\usage{
fvs_scores_day( df,
default.names = TRUE,
item.names = list( Q1 = "Q1", Q1A = "Q1A",
                   Q2 = "Q2", Q2A1 = "Q2A1",
                   Q2A2 = "Q2A2", Q3 = "Q3",
                   Q3A = "Q3A", Q4 = "Q4",
                   Q4A = "Q4A", Q5 = "Q5",
                   Q5A = "Q5A", Q6 = "Q6",
                   Q6A = "Q6A", Q7 = "Q7",
                   Q7A = "Q7A", Q8 = "Q8",
                   Q8A = "Q8A", Q9 = "Q9",
                   Q9A = "Q9A", Q10 = "Q10" ) )
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} containing the columns (dietary items) for computing the scores.}

\item{default.names}{A logical. Defaulted to \code{TRUE} and establishes whether default survey item names (see NCI SAS code linked above) are used. If user-specified names are used, set to \code{FALSE} and specify the column names in \code{item.names}.}

\item{item.names}{A named \code{list} containing the user-specified column names (character vectors) for the survey items in the \code{df}. Ignored if \code{default.names} is \code{TRUE}. Must follow format used in \code{usage}.}
}
\value{
Object of class \code{data.frame} containing the original user-supplied data with the
MyPyramid cup equivalents and MyPyramid servings of fruit & vegetable intake. Column names and descriptions
are as follows:

`frt.veg.ce`: Estimated daily average MyPyramid cup equivalents of fruits & vegetables
`frt.veg.ps`: Estimated daily average MyPyramid servings of fruits & vegetables
}
\description{
Calculate MyPyramid cup equivalents and MyPyramid servings of fruit & vegetable intake
on data collected with the National Cancer Institute's Fruit & Vegetable Intake All-Day Screener in the
EATS. For an example of how the data should be structured prior to feeding it to the function, see
\code{data( fv.data.day )} and \code{help( data( fv.data.day ) )}.
}
\details{
Implements the scoring procedures for data obtained from the National Cancer Institute (NCI)
Fruit & Vegetable Intake All-Day Screener from the EATS. MyPyramid cup equivalents and MyPyramid
servings of fruit & vegetable intake. For a detailed description of the screener, please refer
to the NCI's documentation (see below).

Citation:
Thompson FE, Subar AF, Smith AF, et al. Fruit and vegetable assessment: performance of 2 new
short instruments and a food frequency questionnaire. J Am Diet Assoc. 2002;102(12):1764-1772.
doi:10.1016/s0002-8223(02)90379-2
}
\examples{
library( NCIdietscores )

# using default diet item names

fvs_scores_day( fv.data.day )

# user-specified diet item names but using default names in `item.names`

fvs_scores_day( fv.data.day, default.names = FALSE )

# user specified names

d.user <- setNames( fv.data.day,
                    c( "juice", "veg.mix", "juice.size", "fruit",
                       "fruit.size.a", "fruit.size.b", "salad", "salad.size", "frfy", "frfy.size",
                       "oth.pot", "oth.pot.size", "beans", "beans.size", "oth.veg", "oth.veg.size",
                       "tom", "tom.size", "veg.soup", "veg.soup.size" ) )

# run `fvs_scores_day` without specifying column names, throws error
\dontrun{

  fvs_scores_day( df = d.user, default.names = FALSE )

}


# run `fvs_scores_day`  specifying column names in incorrect format, error thrown

\dontrun{
  cls.list <- list( "juice", "veg.mix", "juice.size", "fruit",
                    "fruit.size.a", "fruit.size.b", "salad", "salad.size", "frfy", "frfy.size",
                    "oth.pot", "oth.pot.size", "beans", "beans.size", "oth.veg", "oth.veg.size",
                    "tom", "tom.size", "veg.soup", "veg.soup.size" )

  fvs_scores_day( df = d.user,
                  default.names = FALSE,
                  item.names = cls.list )
}


# run `fvs_scores_day`  specifying column names, no error

cls.list <- list( Q1 = "juice", Q1A = "juice.size",
                  Q2 = "fruit", Q2A1 = "fruit.size.a",
                  Q2A2 = "fruit.size.b", Q3 = "salad",
                  Q3A = "salad.size", Q4 = "frfy",
                  Q4A = "frfy.size", Q5 = "oth.pot",
                  Q5A = "oth.pot.size", Q6 = "beans",
                  Q6A = "beans.size", Q7 = "oth.veg",
                  Q7A = "oth.veg.size", Q8 = "tom",
                  Q8A = "tom.size", Q9 = "veg.soup",
                  Q9A = "veg.soup.size", Q10 = "veg.mix" )

fvs_scores_day( df = d.user,
                default.names = FALSE,
                item.names = cls.list )


# Set "M" and "E" entries to missing before using function

fv.data.day.me <- fv.data.day
fv.data.day.me[ fv.data.day.me == "M" ] <- NA
fv.data.day.me[ fv.data.day.me == "E" ] <- NA

fvs_scores_day( fv.data.day.me )



## more errors: ##

# incorrect data types
\dontrun{

  fvs_scores_day( df = list( fv.data.day ) )

}

# incorrect formatting of data frequencies
\dontrun{
  fv.data.day.format <- fv.data.day

  nms <- paste0( "Q", 1:10 )

  fv.data.day.format[nms][ fv.data.day.format[nms] == 0 ] <- "Never"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 1 ] <- "1-3 times last month"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 3 ] <- "1-2 times per week"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 2 ] <- "3-4 times per week"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 4 ] <- "5-6 times per week"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 5 ] <- "1 time per day"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 6 ] <- "2 times per day"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 7 ] <- "3 times per day"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 8 ] <- "4 times per day"
  fv.data.day.format[nms][ fv.data.day.format[nms] == 9 ] <- "5 or more times per day"

  fvs_scores_day( df = fv.data.day.format )
}

}
\seealso{
\itemize{
\item \href{https://epi.grants.cancer.gov/diet/screeners/fruitveg/}{Screener Documentation}
\item \href{https://epi.grants.cancer.gov/diet/screeners/fruitveg/scoring/}{Scoring Procedures}
\item \href{https://epi.grants.cancer.gov/diet/shortreg/instruments/eats_all-day.pdf}{The Screener}
\item \href{https://epi.grants.cancer.gov/diet/screeners/sas-program-eats-allday.zip}{Original SAS Code from the NCI}
}
}
