% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvs_scores_meal.R
\name{fvs_scores_meal}
\alias{fvs_scores_meal}
\title{Scores for the Fruit & Vegetable Intake By-Meal Screener in the Eating at America's Table Study (EATS)}
\usage{
fvs_scores_meal( df,
default.names = TRUE,
item.names = list( Q1 = "Q1", Q1A = "Q1A",
                  Q2 = "Q2", Q2A = "Q2A",
                  Q3 = "Q3", Q3A = "Q3A",
                  Q4 = "Q4", Q4A = "Q4A",
                  Q5 = "Q5", Q5A = "Q5A",
                  Q6 = "Q6", Q6A1 = "Q6A1",
                  Q6A2 = "Q6A2", Q7 = "Q7",
                  Q7A = "Q7A", Q8 = "Q8",
                  Q8A1 = "Q8A1", Q8A2 = "Q8A2",
                  Q9 = "Q9", Q9A = "Q9A",
                  Q10 = "Q10", Q10A1 = "Q10A1",
                  Q10A2 = "Q10A2", Q11 = "Q11",
                  Q11A = "Q11A", Q12 = "Q12",
                  Q12A = "Q12A", Q13 = "Q13",
                  Q13A = "Q13A", Q14 = "Q14" ) )
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} containing the columns (dietary items) for computing the scores.}

\item{default.names}{A logical. Defaulted to \code{TRUE} and establishes whether default survey item names (see NCI SAS code linked above) are used. If user-specified names are used, set to \code{FALSE} and specify the column names in \code{item.names}.}

\item{item.names}{A named \code{list} containing the user-specified column names (character vectors) for the survey items in the \code{df}. Ignored if \code{default.names} is \code{TRUE}. Must follow format used in \code{usage}.}
}
\value{
Object of class \code{data.frame} containing the original user-supplied data with the
MyPyramid cup equivalents and MyPyramid servings of fruit & vegetable intake. Column names and descriptions
are as follows:

`frt.veg.ce`: Estimated daily average MyPyramid cup equivalents of fruits & vegetables
`frt.veg.ps`: Estimated daily average MyPyramid servings of fruits & vegetables
}
\description{
Calculate MyPyramid cup equivalents and MyPyramid servings of fruit & vegetable intake
on data collected with the National Cancer Institute's Fruit & Vegetable Intake By-Meal Screener in the
EATS. For an example of how the data should be structured prior to feeding it to the function, see
\code{data( fv.data.meal )} and \code{help( data( fv.data.meal ) )}.
}
\details{
Implements the scoring procedures for data obtained from the National Cancer Institute (NCI)
Fruit & Vegetable Intake By-Meal Screener from the EATS. MyPyramid cup equivalents and MyPyramid
servings of fruit & vegetable intake. For a detailed description of the screener, please refer
to the NCI's documentation (see below).

Citation:
Thompson FE, Subar AF, Smith AF, et al. Fruit and vegetable assessment: performance of 2 new
short instruments and a food frequency questionnaire. J Am Diet Assoc. 2002;102(12):1764-1772.
doi:10.1016/s0002-8223(02)90379-2
}
\examples{
library( NCIdietscores )

# using default diet item names

fvs_scores_meal( fv.data.meal )

# user-specified diet item names but using default names in `item.names`

fvs_scores_meal( fv.data.meal, default.names = FALSE )

# user specified names

d.user <- setNames( fv.data.meal,
                    c( "juice", "fruit.pm", "fruit.pm.size.a", "fruit.pm.size.b",
                       "veg.pm", "veg.pm.size", "tom", "tom.size", "veg.soup", "veg.soup.size",
                       "veg.mix", "juice.size", "salad", "salad.size", "frfy", "frfy.size",
                       "oth.pot", "oth.pot.size", "beans", "beans.size", "fruit.am",
                       "fruit.am.size.a", "fruit.am.size.b", "veg.am", "veg.am.size",
                       "fruit.nn", "fruit.nn.size.a", "fruit.n.size.b", "veg.nn", "veg.nn.size" ) )

# run `fvs_scores_meal` without specifying column names, throws error
\dontrun{

  fvs_scores_meal( df = d.user, default.names = FALSE )

}


# run `fvs_scores_meal`  specifying column names in incorrect format, error thrown

\dontrun{
  cls.list <- list( "juice", "fruit.pm", "fruit.pm.size.a", "fruit.pm.size.b",
                    "veg.pm", "veg.pm.size", "tom", "tom.size", "veg.soup", "veg.soup.size",
                    "veg.mix", "juice.size", "salad", "salad.size", "frfy", "frfy.size",
                    "oth.pot", "oth.pot.size", "beans", "beans.size", "fruit.am",
                    "fruit.am.size.a", "fruit.am.size.b", "veg.am", "veg.am.size",
                    "fruit.nn", "fruit.nn.size.a", "fruit.n.size.b", "veg.nn", "veg.nn.size" )

  fvs_scores_meal( df = d.user,
                   default.names = FALSE,
                   item.names = cls.list )
}


# run `fvs_scores_meal`  specifying column names, no error

cls.list <- list( Q1 = "juice", Q1A = "juice.size",
                  Q2 = "salad", Q2A = "salad.size",
                  Q3 = "frfy", Q3A = "frfy.size",
                  Q4 = "oth.pot", Q4A = "oth.pot.size",
                  Q5 = "beans", Q5A = "beans.size",
                  Q6 = "fruit.am", Q6A1 = "fruit.am.size.a",
                  Q6A2 = "fruit.am.size.b", Q7 = "veg.am",
                  Q7A = "veg.am.size", Q8 = "fruit.nn",
                  Q8A1 = "fruit.nn.size.a", Q8A2 = "fruit.n.size.b",
                  Q9 = "veg.nn", Q9A = "veg.nn.size",
                  Q10 = "fruit.pm", Q10A1 = "fruit.pm.size.a",
                  Q10A2 = "fruit.pm.size.b", Q11 = "veg.pm",
                  Q11A = "veg.pm.size", Q12 = "tom",
                  Q12A = "tom.size", Q13 = "veg.soup",
                  Q13A = "veg.soup.size", Q14 = "veg.mix" )

fvs_scores_meal( df = d.user,
                 default.names = FALSE,
                 item.names = cls.list )


# Set "M" and "E" entries to missing before using function

fv.data.meal.me <- fv.data.meal
fv.data.meal.me[ fv.data.meal.me == "M" ] <- NA
fv.data.meal.me[ fv.data.meal.me == "E" ] <- NA

fvs_scores_meal( fv.data.meal.me )



## more errors: ##

# incorrect data types
\dontrun{

  fvs_scores_meal( df = list( fv.data.meal ) )

}

# incorrect formatting of data frequencies
\dontrun{
  fv.data.meal.format <- fv.data.meal

  nms <- paste0( "Q", 1:14 )

  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 0 ] <- "Never"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 1 ] <- "1-3 times last month"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 3 ] <- "1-2 times per week"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 2 ] <- "3-4 times per week"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 4 ] <- "5-6 times per week"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 5 ] <- "1 time per meal"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 6 ] <- "2 times per meal"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 7 ] <- "3 times per meal"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 8 ] <- "4 times per meal"
  fv.data.meal.format[nms][ fv.data.meal.format[nms] == 9 ] <- "5 or more times per meal"

  fvs_scores_meal( df = fv.data.meal.format )
}
}
\seealso{
\itemize{
\item \href{https://epi.grants.cancer.gov/diet/screeners/fruitveg/}{Screener Documentation}
\item \href{https://epi.grants.cancer.gov/diet/screeners/fruitveg/scoring/bymeal.html}{Scoring Procedures}
\item \href{https://epi.grants.cancer.gov/diet/shortreg/instruments/eats_by-meal.pdf}{The Screener}
\item \href{https://epi.grants.cancer.gov/diet/screeners/sas-program-eats-bymeal.zip}{Original SAS Code from the NCI}
}
}
