% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pef.data.R
\docType{data}
\name{pef.data}
\alias{pef.data}
\title{Test data for examples in the `pef_scores` function.}
\format{
`pef.data`
\describe{
A data frame with 45 rows and 19 columns:
  \item{CEREAL}{Cold cereals freuqnecy of consumption. A value in (1:8)}
  \item{SKIMMILK}{Skim milk freuqnecy of consumption. A value in (1:8)}
  \item{EGGS}{Eggs freuqnecy of consumption. A value in (1:8)}
  \item{SAUSAGE}{Bacon or sausage freuqnecy of consumption. A value in (1:8)}
  \item{MARGBR}{Margarine or butter freuqnecy of consumption. A value in (1:8)}
  \item{CITJUICE}{Orange or grapefruit juice freuqnecy of consumption. A value in (1:8)}
  \item{FRUIT}{Fruit freuqnecy of consumption. A value in (1:8)}
  \item{HOTDOGS}{Beef or pork hotdogs freuqnecy of consumption. A value in (1:8)}
  \item{CHEESE}{Cheese or cheese spread freuqnecy of consumption. A value in (1:8)}
  \item{FRIEDPOT}{French fries, home fries, or hash browns freuqnecy of consumption. A value in (1:8)}
  \item{MARGVEG}{Margarine or butter on vegetables freuqnecy of consumption. A value in (1:8)}
  \item{MAYO}{Mayonnaise freuqnecy of consumption. A value in (1:8)}
  \item{SALDRS}{Salad dressing freuqnecy of consumption. A value in (1:8)}
  \item{RICE}{Rice freuqnecy of consumption. A value in (1:8)}
  \item{MARGRICE}{Margarine, butter, or oil on rice freuqnecy of consumption. A value in (1:8)}
  \item{LOFATMRG}{Reduced fat margarine freuqnecy of consumption. A value in (1:6)}
  \item{ALLFAT}{Subjective fat consumption. A value in (1:3)}
  \item{SEX}{Participant sex, A value in (1,2). 1--males, 2--females. Can also be coded as "male" or "female" for function to work.}
  \item{AGE}{Participant age (years).}
}
}
\source{
<https://epi.grants.cancer.gov/diet/screeners/fat/>
}
\usage{
pef.data
}
\description{
Test data for examples in the `pef_scores` function.
}
\keyword{datasets}
