% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diet.data}
\alias{diet.data}
\title{Test Data for qfs_scores}
\format{

}
\source{
<https://epi.grants.cancer.gov/past-initiatives/open/multifactor/scoring.html>
}
\usage{
diet.data
}
\description{
Test data for examples in the qfs_scores function.
}
\keyword{datasets}
